apps/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
requests==2.31.0
minio==7.2.7
google-cloud-bigquery==3.25.0
